requirements.txt
flask
numpy
pydicom
pyyaml
requests-toolbelt
SimpleITK
torch
torchvision
